Consolidate shared queue tests into templated Catch2 cases.
Replace duplicated MPSCQueue/TaskQueue scenario shells with a single
Queue.cpp using TEMPLATE_TEST_CASE, keeping type-specific stress tests
in their own files and removing queue_bdd_helpers.hpp.
